requirements.txt
coverage==7.4.4
Django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.9